+++ app/controllers/manage/bookings_controller.rb
@@ -67,9 +67,8 @@ module Manage
       @booking.assign_attributes(booking_params)
       process_booking_question_responses
       @booking.save(context: :manage_update)
-      redirect_to = call_booking_action
       write_booking_log
-      respond_with :manage, @booking, location: redirect_to
+      respond_with :manage, @booking
     end
 
     def destroy
@@ -85,7 +84,7 @@ module Manage
     end
 
     def write_booking_log
-      Booking::Log.log(@booking, trigger: :manager, action: booking_action, user: current_user)
+      Booking::Log.log(@booking, trigger: :manager, user: current_user)
     end
 
     def set_filter
@@ -96,18 +95,6 @@ module Manage
       responses = BookingQuestionResponse.process_nested_attributes(@booking, booking_question_responses_params,
                                                                     manage: true)
       @booking.booking_question_responses = responses unless responses.nil?
-    end
-
-    def call_booking_action
-      result = booking_action&.call(booking: @booking)
-      instance_eval(&result.redirect_proc) if result&.redirect_proc.present?
-    rescue BookingActions::Base::NotAllowed
-      @booking.errors.add(:base, :action_not_allowed)
-      nil
-    end
-
-    def booking_action
-      BookingActions::Manage.all[params[:booking_action]]
     end
 
     def booking_params
